Migrate to full Storybook preset based workflow (#14)
This is implemented as as backwards compatible. It is trivial since the backwards compatibility is preset is simply a `presets.js` file requiring `./index.js` 😸 

More background on presets:

- https://storybook.js.org/docs/presets/introduction/
- https://storybook.js.org/docs/presets/writing-presets/ 

The API was made official shortly after `storybook@5` was released. The reasoning for moving to it is that (besides being a cleaner abstraction) it should help with adoptions since we can list `@exemplar` as a preset example here: https://storybook.js.org/docs/presets/preset-gallery/